feat: support action in package.json
