+++ package.json
@@ -47,7 +47,7 @@
 	"devDependencies": {
 		"@commitlint/cli": "11.0.0",
 		"@commitlint/config-conventional": "11.0.0",
-		"@types/chai": "^4.2.14",	
+		"@types/chai": "^4.2.14",
 		"@types/sinon": "^9.0.10",
 		"@types/lodash": "^4.14.165",
 		"@types/mocha": "^8.2.0",
