+++ src/extension.ts
@@ -1,25 +1,17 @@
-import * as vscode from 'vscode';
 import { bas } from './api';
-import { _performAction } from "./actions/performer";
-import { startBasctlServer, closeBasctlServer } from './actions/basctlServer';
+import { startBasctlServer, closeBasctlServer } from './basctlServer/basctlServer';
+import { ActionsController } from './actions/controller';
 
-
-function performScheduledActions() {
-    const actionsSettings = vscode.workspace.getConfiguration();
-    const actionsList: any[] | undefined = actionsSettings.get("actions");
-    if (actionsList && actionsList.length) {
-        for (const action of actionsList) {
-            console.log(`performing action ${action.name} of type ${action.constructor.name}`);
-            _performAction(action);
-        }
-        actionsSettings.update("actions", []);
-    }
-}
+let actionsController;
 
 export async function activate() {
     startBasctlServer();
 
-    performScheduledActions();
+    actionsController = new ActionsController();
+    
+    actionsController.loadActions();
+
+    actionsController.performScheduledActions();
 
     return bas;
 }
